Changed SqlCreateAttribute to SqlTableAttribute.
