+++ babel/src/main/codegen/config.fmpp
@@ -33,26 +33,26 @@ data: {
       "org.apache.calcite.sql.babel.SqlSetTimeZone",
       "org.apache.calcite.sql.babel.OnCommitType",
       "org.apache.calcite.sql.babel.SetType",
-      "org.apache.calcite.sql.babel.SqlCreateAttribute",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeBlockCompression",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeBlockCompression.BlockCompressionOption",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeChecksum",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeChecksum.ChecksumEnabled",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeDataBlockSize",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeDataBlockSize.DataBlockModifier",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeDataBlockSize.DataBlockUnitSize",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeFallback",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeFreeSpace",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeIsolatedLoading",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeIsolatedLoading.OperationLevel",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeJournal",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeJournal.JournalType",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeJournal.JournalModifier",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeJournalTable",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeLog",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeMap",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeMergeBlockRatio",
-      "org.apache.calcite.sql.babel.SqlCreateAttributeMergeBlockRatio.MergeBlockRatioModifier",
+      "org.apache.calcite.sql.babel.SqlTableAttribute",
+      "org.apache.calcite.sql.babel.SqlTableAttributeBlockCompression",
+      "org.apache.calcite.sql.babel.SqlTableAttributeBlockCompression.BlockCompressionOption",
+      "org.apache.calcite.sql.babel.SqlTableAttributeChecksum",
+      "org.apache.calcite.sql.babel.SqlTableAttributeChecksum.ChecksumEnabled",
+      "org.apache.calcite.sql.babel.SqlTableAttributeDataBlockSize",
+      "org.apache.calcite.sql.babel.SqlTableAttributeDataBlockSize.DataBlockModifier",
+      "org.apache.calcite.sql.babel.SqlTableAttributeDataBlockSize.DataBlockUnitSize",
+      "org.apache.calcite.sql.babel.SqlTableAttributeFallback",
+      "org.apache.calcite.sql.babel.SqlTableAttributeFreeSpace",
+      "org.apache.calcite.sql.babel.SqlTableAttributeIsolatedLoading",
+      "org.apache.calcite.sql.babel.SqlTableAttributeIsolatedLoading.OperationLevel",
+      "org.apache.calcite.sql.babel.SqlTableAttributeJournal",
+      "org.apache.calcite.sql.babel.SqlTableAttributeJournal.JournalType",
+      "org.apache.calcite.sql.babel.SqlTableAttributeJournal.JournalModifier",
+      "org.apache.calcite.sql.babel.SqlTableAttributeJournalTable",
+      "org.apache.calcite.sql.babel.SqlTableAttributeLog",
+      "org.apache.calcite.sql.babel.SqlTableAttributeMap",
+      "org.apache.calcite.sql.babel.SqlTableAttributeMergeBlockRatio",
+      "org.apache.calcite.sql.babel.SqlTableAttributeMergeBlockRatio.MergeBlockRatioModifier",
       "org.apache.calcite.sql.babel.SqlCreateTable",
       "org.apache.calcite.sql.ddl.SqlDdlNodes",
       "org.apache.calcite.sql.babel.SqlExecMacro",
