+++ babel/src/main/java/org/apache/calcite/sql/babel/SqlCreateTable.java
@@ -40,7 +40,7 @@ public class SqlCreateTable extends SqlCreate
   public final SqlIdentifier name;
   public final SetType setType;
   public final Volatility volatility;
-  public final List<SqlCreateAttribute> tableAttributes;
+  public final List<SqlTableAttribute> tableAttributes;
   public final SqlNodeList columnList;
   public final SqlNode query;
   public final WithDataType withData;
@@ -52,7 +52,7 @@ public class SqlCreateTable extends SqlCreate
 
   public SqlCreateTable(SqlParserPos pos, SqlCreateSpecifier createSpecifier,
       SetType setType, Volatility volatility, boolean ifNotExists,
-      SqlIdentifier name, List<SqlCreateAttribute> tableAttributes,
+      SqlIdentifier name, List<SqlTableAttribute> tableAttributes,
       SqlNodeList columnList, SqlNode query, WithDataType withData,
       SqlPrimaryIndex primaryIndex, OnCommitType onCommitType) {
     super(OPERATOR, pos, createSpecifier, ifNotExists);
@@ -100,7 +100,7 @@ public class SqlCreateTable extends SqlCreate
     name.unparse(writer, leftPrec, rightPrec);
     if (tableAttributes != null) {
       SqlWriter.Frame frame = writer.startList("", "");
-      for (SqlCreateAttribute a : tableAttributes) {
+      for (SqlTableAttribute a : tableAttributes) {
         writer.sep(",", true);
         a.unparse(writer, 0, 0);
       }
