+++ babel/src/main/java/org/apache/calcite/sql/babel/SqlTableAttribute.java
@@ -20,34 +20,28 @@ import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.parser.SqlParserPos;
 
 /**
- * A <code>SqlCreateAttributeFallback</code> is a CREATE TABLE option
- * for the FALLBACK keyword.
+ * A <code>SqlTableAttribute</code> is a base class that can be used
+ * to create custom options for the SQL CREATE TABLE function.
+ *
+ * <p>To customize table option unparsing, override the method
+ * {@link #unparse(SqlWriter, int, int)}.
  */
-public class SqlCreateAttributeFallback extends SqlCreateAttribute {
-
-  private final boolean no;
-  private final boolean protection;
+public abstract class SqlTableAttribute {
+  private final SqlParserPos pos;
 
   /**
-   * Creates a {@code SqlCreateAttributeFallback}.
+   * Creates a {@code SqlCreateOption}.
    *
-   * @param no  Optional NO keyword
-   * @param protection Optional PROTECTION keyword
    * @param pos  Parser position, must not be null
    */
-  public SqlCreateAttributeFallback(boolean no, boolean protection, SqlParserPos pos) {
-    super(pos);
-    this.no = no;
-    this.protection = protection;
+  public SqlTableAttribute(SqlParserPos pos) {
+    this.pos = pos;
   }
 
-  @Override public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
-    if (no) {
-      writer.keyword("NO");
-    }
-    writer.keyword("FALLBACK");
-    if (protection) {
-      writer.keyword("PROTECTION");
-    }
+  /** Writes a SQL representation of this table option to a writer. */
+  public abstract void unparse(SqlWriter writer, int leftPrec, int rightPrec);
+
+  public SqlParserPos getParserPos() {
+    return pos;
   }
 }
