+++ babel/src/main/java/org/apache/calcite/sql/babel/SqlTableAttributeJournalTable.java
@@ -16,32 +16,34 @@
  */
 package org.apache.calcite.sql.babel;
 
+import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.parser.SqlParserPos;
 
 /**
- * A <code>SqlCreateAttributeLog</code> is a CREATE TABLE option
- * for the LOG attribute.
+ * A <code>SqlTableAttributeJournalTable</code> is a CREATE TABLE option
+ * for the WITH JOURNAL TABLE attribute.
  */
-public class SqlCreateAttributeLog extends SqlCreateAttribute {
+public class SqlTableAttributeJournalTable extends SqlTableAttribute {
 
-  private final boolean loggingEnabled;
+  private final SqlIdentifier tableName;
 
   /**
-   * Creates a {@code SqlCreateAttributeLog}.
+   * Creates a {@code SqlTableAttributeJournalTable}.
    *
-   * @param loggingEnabled  Transient journal logging is enabled
+   * @param tableName  Name of the permanent journal table to be used
    * @param pos  Parser position, must not be null
    */
-  public SqlCreateAttributeLog(boolean loggingEnabled, SqlParserPos pos) {
+  public SqlTableAttributeJournalTable(SqlIdentifier tableName, SqlParserPos pos) {
     super(pos);
-    this.loggingEnabled = loggingEnabled;
+    this.tableName = tableName;
   }
 
-  @Override public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
-    if (!loggingEnabled) {
-      writer.keyword("NO");
-    }
-    writer.keyword("LOG");
+  @Override public void unparse(final SqlWriter writer, final int leftPrec, final int rightPrec) {
+    writer.keyword("WITH");
+    writer.keyword("JOURNAL");
+    writer.keyword("TABLE");
+    writer.sep("=");
+    tableName.unparse(writer, 0, 0);
   }
 }
