+++ babel/src/main/java/org/apache/calcite/sql/babel/SqlTableAttributeLog.java
@@ -16,32 +16,32 @@
  */
 package org.apache.calcite.sql.babel;
 
-import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.parser.SqlParserPos;
 
 /**
- * A <code>SqlCreateAttributeMap</code> is a CREATE TABLE option
- * for the MAP attribute.
+ * A <code>SqlTableAttributeLog</code> is a CREATE TABLE option
+ * for the LOG attribute.
  */
-public class SqlCreateAttributeMap extends SqlCreateAttribute {
+public class SqlTableAttributeLog extends SqlTableAttribute {
 
-  private final SqlIdentifier mapName;
+  private final boolean loggingEnabled;
 
   /**
-   * Creates a {@code SqlCreateAttributeMap}.
+   * Creates a {@code SqlTableAttributeLog}.
    *
-   * @param mapName  Name of an existing contiguous map
+   * @param loggingEnabled  Transient journal logging is enabled
    * @param pos  Parser position, must not be null
    */
-  public SqlCreateAttributeMap(SqlIdentifier mapName, SqlParserPos pos) {
+  public SqlTableAttributeLog(boolean loggingEnabled, SqlParserPos pos) {
     super(pos);
-    this.mapName = mapName;
+    this.loggingEnabled = loggingEnabled;
   }
 
-  @Override public void unparse(final SqlWriter writer, final int leftPrec, final int rightPrec) {
-    writer.keyword("MAP");
-    writer.sep("=");
-    mapName.unparse(writer, 0, 0);
+  @Override public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
+    if (!loggingEnabled) {
+      writer.keyword("NO");
+    }
+    writer.keyword("LOG");
   }
 }
